backend/requirements.txt
absl-py==0.9.0
alabaster==0.7.12
aniso8601==8.0.0
astor==0.8.1
Babel==2.8.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cycler==0.10.0
docutils==0.16
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-WTF==0.14.3
gast==0.2.2
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
idna==2.9
imagesize==1.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.1.3
numpy==1.18.1
numpydoc==0.9.2
oauthlib==3.1.0
opencv-contrib-python==4.1.0.25
opencv-python==4.1.2.30
opt-einsum==3.2.1
packaging==20.3
Pillow==7.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.14.0
snowballstemmer==2.0.0
Sphinx==3.0.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
# tensorflow==2.1.0
termcolor==1.1.0
urllib3==1.25.9
Werkzeug==0.15.4
wrapt==1.12.1
WTForms==2.3.1
Flask-Cors==3.0.8
